# jualdanbeli-STANDALONE/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: jualdanbeli Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List orders for current user
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [buyer, seller]
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order from cart or direct purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created with escrow payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (confirm, ship, deliver, complete, cancel)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/confirm-received:
    post:
      operationId: confirmOrderReceived
      tags: [orders]
      summary: Buyer confirms receipt — triggers escrow release
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt confirmed, payment released to seller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/dispute:
    post:
      operationId: openDispute
      tags: [orders]
      summary: Open a dispute for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeInput"
      responses:
        "201":
          description: Dispute opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── PAYMENTS / ESCROW ──────────────────────────────────────────
...
  /payments/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [payments]
      summary: Request withdrawal from wallet to bank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # ─── SHIPPING ───────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ─── AUTH ──────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, name, role, createdAt, stats]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        stats:
          $ref: "#/components/schemas/UserStats"
        sellerInfo:
          $ref: "#/components/schemas/SellerInfo"
    UserStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalSales:
          type: integer
        averageRating:
          type: number
        totalReviews:
          type: integer
      required: [totalProducts, totalSales, averageRating, totalReviews]
    SellerInfo:
      type: object
      properties:
        shopName:
          type: ["string", "null"]
        shopDescription:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        isVerified:
          type: boolean
      required: [isVerified]
...
    Order:
      type: object
      required: [id, buyerId, sellerId, status, items, totalAmount, shippingCost, escrowStatus, createdAt]
      properties:
        id:
          type: integer
        buyerId:
          type: integer
        buyer:
          $ref: "#/components/schemas/UserProfile"
        sellerId:
          type: integer
        seller:
          $ref: "#/components/schemas/UserProfile"
        status:
          type: string
          enum: [pending_payment, paid, confirmed, shipped, delivered, completed, cancelled, disputed]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        shippingCost:
          type: number
        shippingAddress:
          type: string
        courierId:
          type: ["integer", "null"]
        trackingNumber:
          type: ["string", "null"]
        escrowStatus:
          type: string
          enum: [holding, released, refunded]
        notes:
          type: ["string", "null"]
        dispute:
          $ref: "#/components/schemas/Dispute"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
    OrderInput:
      type: object
      required: [items, shippingAddress, courierId]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        shippingAddress:
          type: string
        courierId:
          type: integer
        notes:
          type: string
        fromCart:
          type: boolean
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        trackingNumber:
          type: string
        notes:
          type: string

    # ─── DISPUTES ──────────────────────────────────────────────────
    Dispute:
      type: object
      required: [id, orderId, reason, status, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        reason:
          type: string
        description:
          type: string
        disputeType:
          type: string
        courierName:
          type: ["string", "null"]
        trackingNumber:
          type: ["string", "null"]
        escalatedToCourier:
          type: ["string", "null"]
        escalatedAt:
          type: ["string", "null"]
          format: date-time
        insuranceClaimStatus:
          type: ["string", "null"]
        insuranceClaimId:
          type: ["string", "null"]
        insuranceAmount:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, investigating, resolved_buyer, resolved_seller, closed]
        ruling:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    DisputeInput:
      type: object
      required: [reason, description]
      properties:
        reason:
          type: string
        description:
          type: string
        disputeType:
          type: string
          enum: [paket_tidak_sampai, paket_hilang, kurir_nakal, paket_rusak, barang_tidak_sesuai, other]
        courierName:
          type: string
        trackingNumber:
          type: string
...
    Transaction:
      type: object
      required: [id, type, amount, status, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [sale, withdrawal, refund, topup]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, rejected]
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    WithdrawalInput:
      type: object
      required: [amount, bankName, accountNumber, accountName]
      properties:
        amount:
          type: number
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
...